If there were more than ~64k blocks then LB computation overflowed, as the number of blocks is stored as a uint. Now number of blocks is first cast to double, as the whole computation is supposed to be done in those units.

+++ grid.cpp
@@ -239,8 +239,8 @@ void balanceLoad(dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid){
       //weight set 
       mpiGrid[cells[i]]->parameters[CellParams::LBWEIGHTCOUNTER]=
          Parameters::loadBalanceGamma +
-         mpiGrid[cells[i]]->number_of_blocks * Parameters::loadBalanceAlpha + 
-         mpiGrid[cells[i]]->number_of_blocks * mpiGrid[cells[i]]->number_of_blocks * Parameters::loadBalanceBeta;
+         (double)(mpiGrid[cells[i]]->number_of_blocks) * Parameters::loadBalanceAlpha + 
+         (double)(mpiGrid[cells[i]]->number_of_blocks) * (double)(mpiGrid[cells[i]]->number_of_blocks) * Parameters::loadBalanceBeta;
       mpiGrid.set_cell_weight(cells[i], mpiGrid[cells[i]]->parameters[CellParams::LBWEIGHTCOUNTER]);
    }
    phiprof::start("dccrg.initialize_balance_load");
